Uplift find_const_ty_from_env
